Refactoring: int to auto.

+++ src/include/matrix_helper.h
@@ -19,16 +19,14 @@ std::vector < std::vector <bmp_real> > inverse(std::vector < std::vector <bmp_re
 class CMatrix {
 
 public:
-   // CMatrix();
-   // ~CMatrix();
 
-    void fill_matrix(const int N_base, matrix_type::_vector z, 
-                     matrix_type::_vector y0, matrix_type::_vector &B, 
+    void fill_matrix(const size_t N_base, matrix_type::_vector z,
+                     matrix_type::_vector y0, matrix_type::_vector &B,
                      matrix_type::_matrix &A);
 
-    void find_coefficients(matrix_type::_matrix A_inv, matrix_type::_vector B, 
-                           matrix_type::_vector &a, matrix_type::_vector &b, 
-                           int N);
+    void find_coefficients(matrix_type::_matrix A_inv, matrix_type::_vector B,
+                           matrix_type::_vector &a, matrix_type::_vector &b,
+                           size_t N);
 
 private:
     friend std::ostream& operator << (std::ostream&, CMatrix& a);
@@ -41,4 +39,4 @@ void GetApproxomateValues(matrix_type::_vector &a,
     matrix_type::_vector &I,
     matrix_type::_vector &z,
     matrix_type::_vector &delta_base,
-    matrix_type::_vector &delta_additional, const int N_base);
+    matrix_type::_vector &delta_additional, const size_t N_base);
